+++ src/actions/ShapeActions.js
@@ -1,103 +1,25 @@
-import {Mesh, MeshBuilder} from "babylonjs";
+import {MeshBuilder} from "babylonjs";
 import Immutable from "immutable";
 import {convertShapeProps, hasEntity, getEntity} from "../classes/Helpers";
 
 const shapeCreators = {
-    box(scene, entityID, options) {
-        // const {size, updatable} = options;
-        return new MeshBuilder.CreateBox(entityID, options, scene);
-        // return new Mesh.CreateBox(entityID, size, scene, updatable);
-    },
-
-    sphere(scene, entityID, options) {
-        const {segments, diameter, updatable} = options;
-        return new Mesh.CreateSphere(entityID, segments, diameter, scene, updatable);
-    },
-
-    ground(scene, entityID, options) {
-        const {width, height, subdivisions, updatable} = options;
-        return new Mesh.CreateGround(entityID, width, height, subdivisions, scene, updatable);
-    },
-
-    groundFromHeightMap(scene, entityID, options, callback) {
-        const {
-            heightMap,
-            meshWidth,
-            meshHeight,
-            subdivisions = 250,
-            minHeight,
-            maxHeight
-        } = options;
-
-        return new Mesh.CreateGroundFromHeightMap(
-            entityID,
-            heightMap,
-            meshWidth,
-            meshHeight,
-            subdivisions,
-            minHeight,
-            maxHeight,
-            scene,
-            true, // updatable
-            callback
-        );
-    },
-
-    disc(scene, entityID, options) {
-        const {
-            radius,
-            tessellation,
-            updatable = true,
-            sideOrientation = null
-        } = options;
-
-        return new Mesh.CreateDisc(entityID, radius, tessellation, scene, updatable, sideOrientation);
-    },
-
-    cylinder(scene, entityID, options) {
-        const {
-            height = 1,
-            diameterTop = 1,
-            diameterBottom = 1,
-            tessellation = 30,
-            subdivisions = 6,
-            updatable = true
-        } = options;
-
-        return new Mesh.CreateCylinder(
-            entityID,
-            height,
-            diameterTop,
-            diameterBottom,
-            tessellation,
-            subdivisions,
-            scene,
-            updatable
-        );
-    },
-
-    torus(scene, entityID, options) {
-        const {
-            diameter = 1,
-            thickness = 1,
-            tessellation = 10,
-            updatable = true,
-            sideOrientation = 0
-        } = options;
-
-        return new Mesh.CreateTorus(entityID, diameter, thickness, tessellation, scene, updatable, sideOrientation);
-    },
-
-    lines(scene, entityID, options) {
-        const {vectors} = options;
-
-        return new Mesh.CreateLines(entityID, vectors, scene);
-    },
-
-    dashedLines(scene, entityID, options) {
-        const {vectors, dashSize, gapSize, dashNumber} = options;
-
-        return new Mesh.CreateDashedLines(entityID, vectors, dashSize, gapSize, dashNumber, scene);
+    Box: (scene, entityID, options) => new MeshBuilder.CreateBox(entityID, options, scene),
+    Sphere: (scene, entityID, options) => new MeshBuilder.CreateSphere(entityID, options, scene),
+    IcoSphere: (scene, entityID, options) => new MeshBuilder.CreateIcoSphere(entityID, options, scene),
+    Ribbon: (scene, entityID, options) => new MeshBuilder.CreateRibbon(entityID, options, scene),
+    Cylinder: (scene, entityID, options) => new MeshBuilder.CreateCylinder(entityID, options, scene),
+    Torus: (scene, entityID, options) => new MeshBuilder.CreateTorus(entityID, options, scene),
+    TorusKnot: (scene, entityID, options) => new MeshBuilder.CreateTorusKnot(entityID, options, scene),
+    LineSystem: (scene, entityID, options) => new MeshBuilder.CreateLineSystem(entityID, options, scene),
+    Lines: (scene, entityID, options) => new MeshBuilder.CreateLines(entityID, options, scene),
+    DashedLines: (scene, entityID, options) => new MeshBuilder.CreateDashedLines(entityID, options, scene),
+    Lathe: (scene, entityID, options) => new MeshBuilder.CreateLathe(entityID, options, scene),
+    Ground: (scene, entityID, options) => new MeshBuilder.CreateGround(entityID, options, scene),
+    TiledGround: (scene, entityID, options) => new MeshBuilder.CreateTiledGround(entityID, options, scene),
+    Tube: (scene, entityID, options) => new MeshBuilder.CreateTube(entityID, options, scene),
+    Polyhedron: (scene, entityID, options) => new MeshBuilder.CreatePolyhedron(entityID, options, scene),
+    GroundFromHeightMap: (scene, entityID, options) => {
+        return new MeshBuilder.CreateGroundFromHeightMap(entityID, options.heightMap, options, scene);
     }
 };
 
@@ -115,19 +37,20 @@ const ShapeActions = {
         if (type && shapeCreators[type]) {
             const scene = state().getIn(["entities", sceneID, "entity"]);
             const options = convertShapeProps(props);
-            let shape = shapeCreators[type](scene, entityID, options, () => {/*action callback goes here.*/});
+
+            const shape = shapeCreators[type](scene, entityID, options, () => {/*action callback goes here.*/});
 
             const meshObj = Immutable.Map({
                 id: entityID,
                 entity: shape,
                 type: "mesh",
-                shape: type
+                shape: type.toLowerCase()
             });
 
             return state().setIn(["entities", entityID], meshObj);
         }
 
-        return state();
+        return null;
     },
 
     updateShape(state, actions, entityID, props) {